add explicit int64_t cast
OS X has trouble deciding between an int and int64_t in this case.

+++ hphp/runtime/base/program-functions.cpp
@@ -287,7 +287,7 @@ void register_variable(Array& variables, char *name, const Variant& value,
       if (!index) {
         symtable->append(Array::Create());
         gpc_elements.push_back(uninit_null());
-        gpc_elements.back().assignRef(symtable->lvalAt(symtable->size() - 1));
+        gpc_elements.back().assignRef(symtable->lvalAt((int64_t)symtable->size() - 1));
       } else {
         String key(index, index_len, CopyString);
         Variant v = symtable->rvalAt(key);
